cloud_io: extract remote_api as a bazel module
